split off discord specific reporter stuff
